pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "probeflow"
version = "1.5.0"
description = "ProbeFlow — STM scan browser, processor, and Createc↔Nanonis toolkit. Browse .sxm scans with colormaps, apply line corrections / background subtraction / smoothing / FFT / grain detection, and convert Createc .dat files to Nanonis .sxm + PNG from either a GUI or a bash-friendly CLI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "scipy",
  "pillow",
  "PySide6",
  "matplotlib",
]

[project.optional-dependencies]
dev       = ["pytest"]
omicron   = ["access2thematrix"]
rhk       = ["spym"]
gwyddion  = ["gwyfile"]
all       = ["access2thematrix", "spym", "gwyfile"]

[project.scripts]
probeflow = "probeflow.cli:main"
dat-sxm   = "probeflow.dat_sxm:main"
dat-png   = "probeflow.dat_png:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = ["probeflow", "probeflow.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
